Delete recursive, cut and new interface

+++ controller.py
@@ -2,7 +2,7 @@ import sys, os
 
 class Controller():
 
-    def copyFile(self, file, location):
+    def cp(self, file, location):
         dest = open(location+"/"+self.basename(file), "w+")
         
         if os.path.exists(file) and os.path.isfile(file):
@@ -32,40 +32,76 @@ class Controller():
                                 new_selection.append(item+"/"+basename)
 
                             if not self.copy(new_selection, dest):
-                                return True
+                                return False
                         else:
-                            if not self.copyFile(item, location):
+                            if not self.cp(item, location):
                                 return False
                 return True
         return False
  
-    def cut(self):
-        return
+    def cut(self, selection, location):
+        if os.path.exists(location):
+            if len(selection) > 0:    
+                for item in selection:
+                    if os.path.exists(item):
+                        if os.path.isdir(item):
+                            dest = location+"/"+self.basename(item)
+                            try:
+                                os.mkdir(dest, 0755)                           
+                            except:
+                                return False
 
-    def delete(self, path):
-        if os.path.exists(path):
-            if os.path.isdir(path):
-                for (i, basename) in enumerate(os.listdir(path)):
-                    item = path + "/" + basename
-                    if os.path.isfile(item):
+                            new_selection = []
+                            for basename in os.listdir(item):
+                                new_selection.append(item+"/"+basename)
+
+                            if self.cut(new_selection, dest):
+                                try:
+                                    os.rmdir(item)
+                                except:
+                                    return False
+                            else:
+                                return False
+                        else:
+                            if self.cp(item, location):
+                                try:
+                                    os.unlink(item)
+                                except:
+                                    return False
+                            else:
+                                return False
+                return True
+        return False
+ 
+    def delete(self, selection, location):
+        if len(selection) > 0:
+            for path in selection:
+                if os.path.exists(path):
+                    if os.path.isdir(path):
+                        for src in os.listdir(path):
+                            item = path + "/" + src
+                            if os.path.isdir(item):
+                                new_selection = []
+                                for basename in os.listdir(item):
+                                    new_selection.append(item)
+
+                                if not self.delete(new_selection, item):
+                                    return False
+                            else:
+                                try:
+                                    os.unlink(item)
+                                except:
+                                    return False
                         try:
-                            os.unlink(item)
+                             os.rmdir(path)    
                         except:
                             return False
                     else:
-                        if not self.delete(item):
+                        try:
+                            os.unlink(path)
+                        except:
                             return False
-                try:
-                     os.rmdir(path)    
-                except:
-                    return False
-            else:
-                try:
-                    os.unlink(path)
-                except:
-                    return False
-            
-            return True    
+            return True
         return False
 
     def createFolder(self, folder):
